moving entity type to type_enums
